New test_factorize to new cudf directory
